Refactor: 자동차 경주 우승자 - View / Domain 분리 - RacingGame, isOver() 추가
