+++ src/main/java/racing/domain/move/MoveStrategy.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.domain.move;
 
 public interface MoveStrategy {
