+++ src/main/java/racing/domain/move/RandomMove.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.domain.move;
 
 import java.util.Random;
