+++ src/test/java/racing/domain/model/CarsTest.java
@@ -1,14 +1,11 @@
-package racing;
+package racing.domain.model;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import java.util.Arrays;
 import java.util.List;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
-import org.junit.jupiter.params.provider.NullAndEmptySource;
+import racing.domain.move.MoveOneSpace;
+import racing.domain.move.NonMove;
 
 class CarsTest {
