+++ src/test/java/racing/domain/move/MoveOddCase.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.domain.move;
 
 public class MoveOddCase implements MoveStrategy {
