+++ src/test/java/racing/domain/move/NonMove.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.domain.move;
 
 public class NonMove implements MoveStrategy {
