+++ src/test/java/racing/domain/move/RandomMoveTest.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.domain.move;
 
 import static org.assertj.core.api.Assertions.assertThat;
